Track a cached copy of node meta/state
Previously the code would go out and try and calculate a node meta state from the database, leading to huge subqueries. These suqueries would take upwards of 20 minutes to complete, and when executed would lead to memory issues because of the shear number of objects to be created (each node version). Now the storage tracks an updated version of the meta so that the subquery/join can be much shorter, and the number of objects is limited to 2 per node, instead of potentially thousands per node.

This resolves #79