Reuse the node queue save logic to avoid conflicting data.
This should fix #110. 

What appears to have been happening was voyager had a huge queue of node update (>2500), and that invoked a condition for voyager to save the node queue. This queue was not correctly sanitized however, and resulted in bad data being saved. This should remove the bad saving code and correctly store the node graph.